Added logic for handling closing of BlockTreeTermsWriter

+++ lucene/index/writer/document.go
@@ -62,7 +62,7 @@ func (w *DocumentWriter) addDocuments(docs []*document.IndexableDocument) error 
 }
 
 func (w *DocumentWriter) flush(sws *segment.SegmentWriteState) error {
-	var bytesWritten int = 0
+	var bytesWritten uint64 = 0
 	n, err := w.term.flush(sws)
 	if err != nil {
 		return err
